op-service: improve beacon client robustness (#17866)
* op-service: always fall back to fetching blob sidecars

Previously, we only reverted to fetching sidecars when an error was
received from the beacon server. In an attempt to be more robust,
we also fetch sidecars when the beacon server gives us a semantically
invalid response.

* op-service: rearrange blobs based on provided hashes

The behavior in the post-Fulu /blobs/ happy path and the
/blob_sidecars/ fallback slightly differed: in the former case,
returned blobs were ordered based their index in the blob, while in the
latter case they were ordered based on the provided hashes.

This bug was not caught in testing because all calls to BeaconBlobs
provide a slice of hashes sorted by index.