op-acceptance-tests: various improvements to the Fusaka acceptance test (#17944)
* split out helpers and initialization to a separate file

* fix comment in mise.toml

* rename osaka test

* rename to fusaka

* txplan: support non-cell proof blob txs

* also verify the l1 blob base fee in the l1 block contract

* optimization

https://github.com/ethereum-optimism/optimism/pull/17529#discussion_r2406241699

* op-acceptance-tests: harden the BPO blob base fee calculation test (#17953)

Before, we merely waited for the blob base fee to rise above 1 before assuming
that the blob parameters would have a visible impact on the fee.

Due to rounding, it is possible that different parameters still result in the same
base fee even above the minimum (either 1 or the reserve price set by EIP-7918).

Calculating the precise intersections between different blob base fee formulae is
complicated, so we opt for the simpler but maximally robust approach: calculate
both the BPO1 base fee and the Osaka base fee, and only proceed when they differ.

This resolves a nit on a previous PR that was merged due to time pressure:
https://github.com/ethereum-optimism/optimism/pull/17529#discussion_r2405498871

* add log