op-service/sources: detect duplicate blob hashes
Update blob fetching logic to handle the case where multiple blobs with the same content (but different indices in the block) are requested. When reordering the blob responses matching was done purely on blob hash which meant some output blobs were written twice and some not set at all. This resulted in an error about nil blobs. The fix is to assign to a blob hash request that hasn't previously been set.

Signed-off-by: pandainzoo <gary.zhu@mantle.xyz>